feat: publish new packages to CDN (#48)
Update the publish package workflow so that new
packages are now also published to the CDN.

Each new package will be published to a versioned directory
in the CDN and replace the files for its @latest version. This
will allow users to either link to versioned assets or the latest files.